Minor refactoring & code restructuring
