+++ src/views/settingsv2/platforms_panel.tsx
