+++ src/views/settingsv2/tab.tsx
@@ -15,7 +15,7 @@ function a11yProps(index: number) {
 }
 
 export default function SettingsTab() {
-  const [value, setValue] = React.useState(1);
+  const [value, setValue] = React.useState(0);
 
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
     setValue(newValue);
